Restructure component and add GA tracking.
