+++ assets/js/modules/reader-revenue-manager/components/settings/SettingsEdit.js
@@ -19,43 +19,38 @@
 /* eslint complexity: [ "error", 17 ] */
 
 /**
+ * External dependencies
+ */
+import { useUnmount } from 'react-use';
+
+/**
  * WordPress dependencies
  */
-import { createInterpolateElement } from '@wordpress/element';
-import { __, sprintf } from '@wordpress/i18n';
+import { useEffect, useState } from '@wordpress/element';
 
 /**
  * Internal dependencies
  */
 import { ProgressBar } from 'googlesitekit-components';
 import { useSelect } from 'googlesitekit-data';
-import { useFeature } from '../../../../hooks/useFeature';
+import useViewContext from '../../../../hooks/useViewContext';
+import { getPostTypesString } from '../../utils/settings';
+import { trackEvent } from '../../../../util';
 import { CORE_MODULES } from '../../../../googlesitekit/modules/datastore/constants';
+import { CORE_SITE } from '../../../../googlesitekit/datastore/site/constants';
 import {
 	MODULES_READER_REVENUE_MANAGER,
 	READER_REVENUE_MANAGER_MODULE_SLUG,
 } from '../../datastore/constants';
-import ErrorText from '../../../../components/ErrorText';
-import {
-	PostTypesSelect,
-	PublicationOnboardingStateNotice,
-	PublicationSelect,
-	SnippetModeSelect,
-} from '../common';
-import SettingsNotice, {
-	TYPE_INFO,
-} from '../../../../components/SettingsNotice';
-import WarningIcon from '../../../../../../assets/svg/icons/warning-icon.svg';
-import StoreErrorNotices from '../../../../components/StoreErrorNotices';
-import ProductIDSettings from './ProductIDSettings';
+import { SNIPPET_MODES } from '../../constants';
+import SettingsForm from './SettingsForm';
 
 export default function SettingsEdit() {
-	const isRRMv2Enabled = useFeature( 'rrmModuleV2' );
+	const viewContext = useViewContext();
 
 	const isDoingSubmitChanges = useSelect( ( select ) =>
 		select( MODULES_READER_REVENUE_MANAGER ).isDoingSubmitChanges()
 	);
-
 	const hasModuleAccess = useSelect( ( select ) => {
 		const { hasModuleOwnershipOrAccess, getErrorForAction } =
 			select( CORE_MODULES );
@@ -88,79 +83,53 @@ export default function SettingsEdit() {
 
 		return false;
 	} );
-
-	const publicationID = useSelect( ( select ) =>
-		select( MODULES_READER_REVENUE_MANAGER ).getPublicationID()
+	const haveSettingsChanged = useSelect( ( select ) =>
+		select( MODULES_READER_REVENUE_MANAGER ).haveSettingsChanged()
+	);
+	const allPostTypes = useSelect( ( select ) =>
+		select( CORE_SITE ).getPostTypes()
+	);
+	const settings = useSelect( ( select ) =>
+		select( MODULES_READER_REVENUE_MANAGER ).getSettings()
 	);
 
-	const missingProductID = useSelect( ( select ) => {
-		const productID = select(
-			MODULES_READER_REVENUE_MANAGER
-		).getProductID();
+	const { snippetMode, postTypes } = settings || {};
 
-		if ( productID === undefined ) {
-			return undefined;
+	const [ oldSettings, setOldSettings ] = useState( settings );
+
+	// Set old settings if they were not available on mount.
+	useEffect( () => {
+		if ( ! oldSettings && !! settings ) {
+			setOldSettings( settings );
+		}
+	}, [ oldSettings, settings ] );
+
+	// Track GA event when snippet mode or post types change.
+	useUnmount( () => {
+		// Do not run if settings have not been saved.
+		if ( haveSettingsChanged ) {
+			return;
 		}
 
-		if ( 'openaccess' === productID ) {
-			return null;
+		if ( snippetMode !== oldSettings.snippetMode ) {
+			trackEvent(
+				`${ viewContext }_rrm-settings`,
+				'change_snippet_mode',
+				SNIPPET_MODES[ snippetMode ]
+			);
 		}
 
-		const productIDs = select(
-			MODULES_READER_REVENUE_MANAGER
-		).getProductIDs();
-
-		if ( productIDs === undefined ) {
-			return undefined;
+		if (
+			getPostTypesString( postTypes ) !==
+			getPostTypesString( oldSettings.postTypes )
+		) {
+			trackEvent(
+				`${ viewContext }_rrm-settings`,
+				'change_post_types',
+				getPostTypesString( postTypes, allPostTypes )
+			);
 		}
-
-		return productIDs.includes( productID ) ? null : productID;
 	} );
-
-	const publicationAvailable = useSelect( ( select ) => {
-		if ( hasModuleAccess === undefined ) {
-			return undefined;
-		}
-
-		if ( hasModuleAccess === false ) {
-			return false;
-		}
-
-		const publications = select(
-			MODULES_READER_REVENUE_MANAGER
-		).getPublications();
-
-		if ( ! Array.isArray( publications ) ) {
-			return undefined;
-		}
-
-		return publications.some(
-			// eslint-disable-next-line sitekit/acronym-case
-			( { publicationId: id } ) => id === publicationID
-		);
-	} );
-	const formattedOwnerName = useSelect( ( select ) => {
-		const module = select( CORE_MODULES ).getModule(
-			READER_REVENUE_MANAGER_MODULE_SLUG
-		);
-
-		return module?.owner?.login
-			? `<strong>${ module.owner.login }</strong>`
-			: __( 'Another admin', 'google-site-kit' );
-	} );
-	const snippetMode = useSelect( ( select ) =>
-		select( MODULES_READER_REVENUE_MANAGER ).getSnippetMode()
-	);
-	const productIDs = useSelect( ( select ) =>
-		select( MODULES_READER_REVENUE_MANAGER ).getProductIDs()
-	);
-	const publicationsLoaded = useSelect(
-		( select ) =>
-			hasModuleAccess === false ||
-			select( MODULES_READER_REVENUE_MANAGER ).hasFinishedResolution(
-				'getPublications'
-			)
-	);
 
 	if ( isDoingSubmitChanges || undefined === hasModuleAccess ) {
 		return <ProgressBar />;
@@ -168,99 +137,7 @@ export default function SettingsEdit() {
 
 	return (
 		<div className="googlesitekit-setup-module googlesitekit-setup-module--reader-revenue-manager googlesitekit-rrm-settings-edit">
-			<div className="googlesitekit-settings-module__fields-group">
-				<StoreErrorNotices
-					moduleSlug={ READER_REVENUE_MANAGER_MODULE_SLUG }
-					storeName={ MODULES_READER_REVENUE_MANAGER }
-				/>
-
-				{ hasModuleAccess && false === publicationAvailable && (
-					<ErrorText
-						message={ sprintf(
-							/* translators: 1: Publication ID. */
-							__(
-								'The previously selected publication with ID %s was not found. Please select a new publication.',
-								'google-site-kit'
-							),
-							publicationID
-						) }
-					/>
-				) }
-
-				{ isRRMv2Enabled &&
-					hasModuleAccess &&
-					publicationAvailable &&
-					missingProductID && (
-						<ErrorText
-							message={ sprintf(
-								/* translators: 1: Product ID. */
-								__(
-									'The previously selected product ID %s was not found. Please select a new product ID.',
-									'google-site-kit'
-								),
-								missingProductID
-							) }
-						/>
-					) }
-
-				<div className="googlesitekit-setup-module__inputs">
-					<PublicationSelect hasModuleAccess={ hasModuleAccess } />
-				</div>
-				{ hasModuleAccess && publicationAvailable && (
-					<PublicationOnboardingStateNotice />
-				) }
-				{ ! hasModuleAccess && (
-					<SettingsNotice
-						type={ TYPE_INFO }
-						Icon={ WarningIcon }
-						notice={ createInterpolateElement(
-							sprintf(
-								/* translators: %s: module owner's name */
-								__(
-									'%s configured Reader Revenue Manager and you don’t have access to its configured publication. Contact them to share access or change the configured publication.',
-									'google-site-kit'
-								),
-								formattedOwnerName
-							),
-							{
-								strong: <strong />,
-							}
-						) }
-					/>
-				) }
-				{ isRRMv2Enabled &&
-					publicationsLoaded &&
-					productIDs?.length > 0 && (
-						<ProductIDSettings
-							hasModuleAccess={ hasModuleAccess }
-						/>
-					) }
-			</div>
-			{ isRRMv2Enabled && (
-				<div className="googlesitekit-settings-module__fields-group">
-					<h4 className="googlesitekit-settings-module__fields-group-title">
-						{ __( 'CTA Placement', 'google-site-kit' ) }
-					</h4>
-					<div className="googlesitekit-rrm-settings-edit__snippet-mode">
-						<SnippetModeSelect
-							hasModuleAccess={ hasModuleAccess }
-						/>
-					</div>
-					{ snippetMode === 'post_types' && (
-						<div className="googlesitekit-rrm-settings-edit__post-types">
-							<h5>
-								{ __(
-									'Select the content types where you want your CTAs to appear:',
-									'google-site-kit'
-								) }
-							</h5>
-							<PostTypesSelect
-								hasModuleAccess={ hasModuleAccess }
-							/>
-						</div>
-					) }
-				</div>
-			) }
+			<SettingsForm hasModuleAccess={ hasModuleAccess } />
 		</div>
 	);
 }
